Reorganized Java classes; support XML Resolver 2.x
